fix: moving get chat messages logic to chat UI
